test: add full unit + functional test suite with CodeCov coverage
- Delete dead Wishlist Application/Domain/Event code (handlers and
  commands were never wired to any controller; replaced by Collection-
  level isWished flag mechanics)
- Add NullMangaApiClient stub for tests (no real HTTP to Jikan/Google)
- Register DAMA DoctrineTestBundle (test env) for transaction rollback
  between functional tests; add DAMA extension to phpunit.dist.xml
- Create config/packages/test/services.yaml overriding external API
  clients with the null stub
- Add .env.test entries (GATE_PASSWORD, NOTIFICATION_EMAIL, etc.)
- Unit tests: Auth/GateUser, all domain exceptions and enums, Manga
  and Volume entities, Collection Entry/VolumeEntry, Notification
  ActivityLog/Notification entities, all domain service value objects,
  ActivityLogEventHandler with mocked repositories
- Functional tests: GateController (valid/invalid/missing password),
  MangaController (CRUD, volume add/update, external search stubs),
  CollectionController (all 12 endpoints, toggle logic, wishlist ops),
  WishlistController (list, add-remaining, clear, purchase),
  StatsController, NotificationController, ArticleController
- Fix Article.php and Notification.php: replace FQCN \DateTimeImmutable
  and \DateTimeInterface with proper use imports (CLAUDE.md rule)
- CI tests.yml: switch to pcov coverage driver, generate clover.xml,
  upload to Codecov via codecov/codecov-action@v5
- Add .codecov.yml: patch + project status checks, PR comment layout

+++ back/src/Collection/Domain/CollectionEntry.php
@@ -70,7 +70,7 @@ class CollectionEntry
             'notificationsEnabled' => $this->notificationsEnabled,
             'notificationsEnabledAt' => $this->notificationsEnabledAt?->format(DateTimeInterface::ATOM),
             'addedAt' => $this->addedAt->format(DateTimeInterface::ATOM),
-            'ownedValue' => array_sum(array_map(
+            'ownedValue' => (float) array_sum(array_map(
                 fn (VolumeEntry $ve) => $ve->isOwned ? ($ve->volume->price ?? 0.0) : 0.0,
                 $this->volumeEntries->toArray(),
             )),
